+++ src/fixtureRepository/index.ts
@@ -1,13 +1,26 @@
-import glob from 'glob';
+import glob from 'fast-glob';
 import { ValueGeneratorBase } from './generators/valueGeneratorBase';
-import { TypeChecker, ContainerItem, Container } from '../typeChecker';
+import { TypeChecker, Container } from '../typeChecker';
+
+export interface FrOptions {
+  ignoreGlob?: string[];
+  additionalFiles?: string[];
+}
+
+const defaultOptions = {
+  ignoreGlob: [],
+  additionalFiles: []
+};
 
 export default class FixtureRepository {
   private static valueGenerator: ValueGeneratorBase;
   private static container: Container;
 
-  public static setup(pattern: string) {
-    const fileNames = glob.sync(pattern);
+  public static setup(pattern: string, options: FrOptions = defaultOptions) {
+    const additionalFiles = options.additionalFiles || defaultOptions.additionalFiles;
+
+    let fileNames = glob.sync(pattern, this.getOptions(options));
+    fileNames = fileNames.concat(additionalFiles);
 
     const typeChecker = new TypeChecker(fileNames);
 
@@ -17,5 +30,13 @@ export default class FixtureRepository {
 
   public static create(type: string): any {
     return this.valueGenerator.resolveAndGenerate(type);
+  }
+
+  private static getOptions(options: FrOptions) {
+    return {
+      onlyFiles: true,
+      unique: true,
+      ignore: options.ignoreGlob || defaultOptions.ignoreGlob
+    };
   }
 }